Improved description of the Coglin RevParry ego
With this fix the description of the ego clearly states that it only works
for melee attacks.

That it indeed only works for melee attacks is clear from the call to
`melee_attack::try_parry_disarm()`, which handles the ego's parry. There
is no corresponding parry-method in `ranged_attack.cc`.